Derive Pitch and Angle from Velocity
Sometimes in HD's codebase, a bullet is not given a pitch that reflects it's current velocity; to alleviate this the pitch and angle can be derived from the velocity vector of the bullet itself. This probably isn't optimized very well for executing thousands of times per frame, but it works well for the moment.
Ideally we would be able to tell whenever a bullet has it's pitch and angle unset, though that doesn't seem to be possible given that the default exists in the superset of possible bullet angles (180). Since the end results are identical, we may be able to just check for 180 after-all, but this ignores bullets that have an incorrect pitch on purpose (not that I've seen any in the wild).